Re-organize packages and fix #290
