+++ simphony/engine/__init__.py
@@ -4,19 +4,12 @@ This module is dynamicaly populated at import with the
 registered plugins modules. Plugins modules need to be
 registered at the 'simphony.engine' entry point.
 """
-from .extension import ABCEngineExtension
-from .extension import EngineInterface
-from .extension import EngineManager
+from ..ext import get_engine_manager
+from ..ext import create_wrapper
 
 
-__all__ = ['ABCEngineExtension', 'EngineInterface',
-           'get_supported_engines', 'create_wrapper',
+__all__ = ['get_supported_engines', 'create_wrapper',
            'get_supported_engine_names']
-
-
-# TODO: Use an application server and put this in app context.
-# Wrapper manager class.
-_ENGINE_MANAGER = EngineManager()
 
 
 def get_supported_engine_names():
@@ -27,7 +20,7 @@ def get_supported_engine_names():
     names: list
         a list of engine names
     """
-    return _ENGINE_MANAGER.get_supported_engine_names()
+    return get_engine_manager().get_supported_engine_names()
 
 
 def get_supported_engines():
@@ -38,27 +31,7 @@ def get_supported_engines():
     metadata: list
         a list of engine metadata objects
     """
-    return _ENGINE_MANAGER.get_supported_engines()
-
-
-def create_wrapper(cuds, engine_name, engine_interface=None):
-    """Create a wrapper to the given engine.
-
-    Parameters
-    ----------
-    cuds: CUDS
-        a cuds object which contains model information
-    engine_name: str
-        name of the underlying engine to launch the simulation with
-    engine_interface: engine.EngineInterface
-        the interface to the engine, internal or fileio
-
-    Returns
-    -------
-    wrapper: engine.ABCEngineExtension
-        an engine wrapper instance
-    """
-    return _ENGINE_MANAGER.create_wrapper(cuds, engine_name, engine_interface)
+    return get_engine_manager().get_supported_engines()
 
 
 def load_engine_extentions():
@@ -70,10 +43,11 @@ def load_engine_extentions():
         namespace='simphony.engine',
         invoke_on_load=False)
     extensions = {}
+    engine_manager = get_engine_manager()
     for ext in mgr.extensions:
         extensions[ext.name] = ext.plugin
         # Load engine metadata
-        _ENGINE_MANAGER.load_metadata(ext.plugin)
+        engine_manager.load_metadata(ext.plugin)
     return extensions
